requirements.txt
Django==1.4.3
MySQL-python==1.2.4c1
South==0.7.6
Werkzeug==0.8.3
amqp==1.0.6
anyjson==0.3.3
argparse==1.2.1
billiard==2.7.3.19
celery==3.0.12
django-celery==3.0.11
django-debug-toolbar==0.9.4
django-dtpanel-htmltidy==0.1.1
django-extensions==1.0.1
gunicorn==0.16.1
kombu==2.5.4
pyserial==2.6
python-dateutil==1.5
pytidylib==0.2.1
pytz==2012h
simplejson==2.6.2
wsgiref==0.1.2
